Astrological Insight Generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0

# Google Gemini (Free tier available)
google-generativeai==0.3.1

# HuggingFace Inference API (Free tier available)
# Uses requests library (already included above)

# Optional: For OpenAI integration (uncomment if using)
# openai==1.3.0

# Optional: For local HuggingFace models (uncomment if using)
# transformers==4.35.0
# torch==2.1.0

# Optional: For Hindi translation (uncomment if using)
# indic-trans==2.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
